# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chat Systems API",
    "description": "Every error response carries `{\"error\": \"<message>\"}` with the HTTP status saying what kind of error it is.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/channels/{ch_id}/messages": {
      "get": {
        "tags": [
          "messages"
        ],
        "operationId": "messaging_list_messages",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "uuid"
            }
          },
          {
            "name": "ch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataList_MessageWithReactions"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "messages"
        ],
        "operationId": "messaging_send_message",
        "parameters": [
          {
            "name": "ch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          }
        }
      }
    },
...
    "/scheduled-messages/{id}": {
      "delete": {
        "tags": [
          "scheduled"
        ],
        "operationId": "cancel_scheduled",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "scheduled"
        ],
        "operationId": "reschedule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RescheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduledMessage"
                }
              }
            }
          }
        }
      }
    },
...
    "/workspaces/{ws_id}/scheduled-messages": {
      "get": {
        "tags": [
          "scheduled"
        ],
        "operationId": "list_scheduled",
        "parameters": [
          {
            "name": "ws_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataList_ScheduledMessage"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "scheduled"
        ],
        "operationId": "create_scheduled",
        "parameters": [
          {
            "name": "ws_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduledMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduledMessage"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BotIdentity": {
        "type": "object",
        "required": [
          "hook_id",
          "name"
        ],
        "properties": {
          "hook_id": {
            "type": "string",
            "format": "uuid"
          },
          "icon_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "CreateScheduledMessageRequest": {
        "type": "object",
        "required": [
          "channel_id",
          "content",
          "send_at"
        ],
        "properties": {
          "channel_id": {
            "type": "string",
            "format": "uuid"
          },
          "content": {
            "type": "string"
          },
          "send_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "DataList_MessageWithReactions": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Message"
                },
                {
                  "type": "object",
                  "required": [
                    "reactions"
                  ],
                  "properties": {
                    "reactions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Reaction"
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      },
...
      "DataList_ScheduledMessage": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "workspace_id",
                "user_id",
                "channel_id",
                "content",
                "send_at",
                "created_at"
              ],
              "properties": {
                "canceled_at": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time"
                },
                "channel_id": {
                  "type": "string",
                  "format": "uuid"
                },
                "content": {
                  "type": "string"
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "failure": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "send_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "sent_at": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time"
                },
                "user_id": {
                  "type": "string",
                  "format": "uuid"
                },
                "workspace_id": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      },
...
      "Message": {
        "type": "object",
        "required": [
          "id",
          "channel_id",
          "user_id",
          "content",
          "metadata",
          "reply_count",
          "is_pinned",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "channel_id": {
            "type": "string",
            "format": "uuid"
          },
          "client_message_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "deleted_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_pinned": {
            "type": "boolean"
          },
          "metadata": {
            "$ref": "#/components/schemas/MessageMetadata"
          },
          "reply_count": {
            "type": "integer",
            "format": "int32"
          },
          "thread_parent_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "MessageMetadata": {
        "type": "object",
        "properties": {
          "bot": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BotIdentity"
              }
            ]
          },
          "huddle_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "initiator_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "kind": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "Reaction": {
        "type": "object",
        "required": [
          "id",
          "message_id",
          "user_id",
          "emoji",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "emoji": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "message_id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "RescheduleRequest": {
        "type": "object",
        "required": [
          "send_at"
        ],
        "properties": {
          "send_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ScheduledMessage": {
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "user_id",
          "channel_id",
          "content",
          "send_at",
          "created_at"
        ],
        "properties": {
          "canceled_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "channel_id": {
            "type": "string",
            "format": "uuid"
          },
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "failure": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "send_at": {
            "type": "string",
            "format": "date-time"
          },
          "sent_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "workspace_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "SendMessageRequest": {
        "type": "object",
        "required": [
          "content"
        ],
        "properties": {
          "client_message_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "The sender's own id for this send, used only to make a retry idempotent.\nIt is not the message id — the server owns that."
          },
          "content": {
            "type": "string"
          },
          "thread_parent_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          }
        }
      },
...
      "StatusResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Same-origin clients are authenticated by the httpOnly cookie instead."
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
...
}